Moved lots of controllers/code to tidy it up into more logical groups

+++ routes/web.php
@@ -34,10 +34,7 @@ Route::group(['middleware' => 'auth'], function () {
     Route::post('/choices', 'ChoiceController@store')->name('projects.choose');
     Route::get('/thank-you', 'ChoiceController@thankYou')->name('thank_you');
 
-    Route::get('/myprofile', 'ProfileController@edit')->name('profile.edit');
-    Route::post('/myprofile', 'ProfileController@update')->name('profile.update');
-
-    Route::delete('/impersonate', 'ImpersonationController@destroy')->name('impersonate.stop');
+    Route::delete('/impersonate', 'Admin\ImpersonationController@destroy')->name('impersonate.stop');
 
     Route::group(['middleware' => 'admin', 'prefix' => '/admin'], function () {
 
@@ -48,28 +45,28 @@ Route::group(['middleware' => 'auth'], function () {
         Route::get('/projects/{category}', 'Admin\ProjectController@index')->name('admin.project.index');
         Route::get('/choices/{category}', 'Admin\ChoiceController@index')->name('admin.student.choices');
 
-        Route::get('/course', 'CourseController@index')->name('admin.course.index');
-        Route::get('/course/create', 'CourseController@create')->name('admin.course.create');
-        Route::get('/course/{course}', 'CourseController@show')->name('admin.course.show');
-        Route::post('/course', 'CourseController@store')->name('admin.course.store');
-        Route::get('/course/{course}/edit', 'CourseController@edit')->name('admin.course.edit');
-        Route::post('/course/{course}', 'CourseController@update')->name('admin.course.update');
-        Route::delete('/course/{course}', 'CourseController@destroy')->name('admin.course.destroy');
+        Route::get('/course', 'Admin\CourseController@index')->name('admin.course.index');
+        Route::get('/course/create', 'Admin\CourseController@create')->name('admin.course.create');
+        Route::get('/course/{course}', 'Admin\CourseController@show')->name('admin.course.show');
+        Route::post('/course', 'Admin\CourseController@store')->name('admin.course.store');
+        Route::get('/course/{course}/edit', 'Admin\CourseController@edit')->name('admin.course.edit');
+        Route::post('/course/{course}', 'Admin\CourseController@update')->name('admin.course.update');
+        Route::delete('/course/{course}', 'Admin\CourseController@destroy')->name('admin.course.destroy');
 
         Route::get('/course/{course}/enroll', 'Admin\EnrollmentController@create')->name('admin.course.enrollment');
         Route::post('/course/{course}/enrollment', 'Admin\EnrollmentController@store')->name('admin.course.enroll');
+        Route::delete('/course/{id}/remove-students', 'Admin\EnrollmentController@destroy')->name('course.remove_students');
 
-        Route::get('/programme', 'ProgrammeController@index')->name('admin.programme.index');
-        Route::get('/programme/create', 'ProgrammeController@create')->name('admin.programme.create');
-        Route::post('/programme', 'ProgrammeController@store')->name('admin.programme.store');
-        Route::get('/programme/{id}', 'ProgrammeController@edit')->name('admin.programme.edit');
-        Route::post('/programme/{id}', 'ProgrammeController@update')->name('admin.programme.update');
-        Route::delete('/programme/{id}', 'ProgrammeController@destroy')->name('admin.programme.destroy');
+        Route::get('/programme', 'Admin\ProgrammeController@index')->name('admin.programme.index');
+        Route::get('/programme/create', 'Admin\ProgrammeController@create')->name('admin.programme.create');
+        Route::post('/programme', 'Admin\ProgrammeController@store')->name('admin.programme.store');
+        Route::get('/programme/{id}', 'Admin\ProgrammeController@edit')->name('admin.programme.edit');
+        Route::post('/programme/{id}', 'Admin\ProgrammeController@update')->name('admin.programme.update');
+        Route::delete('/programme/{id}', 'Admin\ProgrammeController@destroy')->name('admin.programme.destroy');
 
-        Route::post('/bulk-accept', 'BulkAcceptanceController@store')->name('project.bulk_accept');
-        Route::delete('/course/{id}/remove-students', 'CourseMemberController@destroy')->name('course.remove_students');
+        Route::post('/bulk-accept', 'Admin\BulkAcceptanceController@store')->name('project.bulk_accept');
 
-        Route::post('/impersonate/{id}', 'ImpersonationController@store')->name('impersonate.start');
+        Route::post('/impersonate/{id}', 'Admin\ImpersonationController@store')->name('impersonate.start');
 
         Route::get('/users/{category}', 'Admin\UserController@index')->name('admin.users');
         Route::get('/user/{user}', 'Admin\UserController@show')->name('admin.user.show');
@@ -77,11 +74,11 @@ Route::group(['middleware' => 'auth'], function () {
 
         Route::post('/user/{user}/toggle-admin', 'Admin\UserController@toggleAdmin')->name('admin.users.toggle_admin');
 
-        Route::delete('/students/remove/undergrad', 'BulkRemovalController@undergrads')->name('students.remove_undergrads');
-        Route::delete('/students/remove/postgrad', 'BulkRemovalController@postgrads')->name('students.remove_postgrads');
-        Route::delete('/students/remove-all', 'BulkRemovalController@all')->name('students.remove_all');
+        Route::delete('/students/remove/undergrad', 'Admin\BulkRemovalController@undergrads')->name('students.remove_undergrads');
+        Route::delete('/students/remove/postgrad', 'Admin\BulkRemovalController@postgrads')->name('students.remove_postgrads');
+        Route::delete('/students/remove-all', 'Admin\BulkRemovalController@all')->name('students.remove_all');
 
-        Route::get('/export/projects-excel', 'ExportController@projects')->name('export.projects.excel');
+        Route::get('/export/projects-excel', 'Admin\ExportController@projects')->name('export.projects.excel');
 
         Route::get('/gdpr/user/{user}', 'Gdpr\UserExportController@show')->name('gdpr.export.user');
